- all min/  max functions operating on containers or with iterators   have as a return type the value_type of the container - further improvements on the MultiSTA RePro.
git-svn-id: https://svn.code.sf.net/p/relacs/code/relacs/trunk@331 1550813e-0d35-0410-aec7-fc86b090456a

+++ plugins/linuxdevices/comedi/src/comedianalogoutput.cc
@@ -449,8 +449,8 @@ int ComediAnalogOutput::setupCommand( OutList &sigs, comedi_cmd &cmd )
     double min = sigs[k].requestedMin();
     double max = sigs[k].requestedMax();
     if ( min == OutData::AutoRange || max == OutData::AutoRange ) {
-      double smin = 0.0;
-      double smax = 0.0;
+      float smin = 0.0;
+      float smax = 0.0;
       minMax( smin, smax, sigs[k] );
       if ( min == OutData::AutoRange )
 	min = smin;
